+++ api-gateway/method-handler/main.tf
@@ -1,3 +1,11 @@
+resource "aws_api_gateway_request_validator" "this" {
+  count = var.request_model_schema != null ? 1 : 0
+
+  rest_api_id           = var.api_id
+  name                  = "${var.http_method}-${var.resource_id}-validator"
+  validate_request_body = true
+}
+
 resource "aws_api_gateway_model" "this" {
   count = var.request_model_schema != null ? 1 : 0
 
@@ -13,13 +21,11 @@ resource "aws_api_gateway_method" "this" {
   http_method          = var.http_method
   authorization        = var.authorization
   authorizer_id        = var.authorizer_id
-  request_validator_id = var.request_validator_id
+  request_validator_id = var.request_model_schema != null ? aws_api_gateway_request_validator.this[0].id : null
 
   request_models = var.request_model_schema != null ? {
     "application/json" = aws_api_gateway_model.this[0].name
-  } : (var.request_model_name != null ? {
-    "application/json" = var.request_model_name
-  } : null)
+  } : null
 }
 
 resource "aws_api_gateway_integration" "this" {